Split import bundle task to a separate plugin
