+++ gateway-connection-api/src/main/java/com/ca/apim/gateway/connection/GatewayClient.java
@@ -4,7 +4,7 @@
  * of the MIT license.  See the LICENSE file for details.
  */
 
-package com.ca.apim.gateway.cagatewayexport.util.http;
+package com.ca.apim.gateway.connection;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.auth.UsernamePasswordCredentials;
