+++ gateway-connection-api/src/main/java/com/ca/apim/gateway/connection/GatewayConnectionProperties.java
@@ -4,7 +4,7 @@
  * of the MIT license.  See the LICENSE file for details.
  */
 
-package com.ca.apim.gateway.cagatewayexport.config;
+package com.ca.apim.gateway.connection;
 
 import org.gradle.api.Project;
 import org.gradle.api.provider.Property;
