+++ gateway-export-plugin/build.gradle
@@ -18,10 +18,7 @@ project.ext.'gradle.publish.secret' = System.env.GRADLE_PUBLISH_SECRET
 dependencies {
     compile gradleApi()
     compile project(':config-builder')
-    compile 'org.jetbrains:annotations:16.0.2'
-    compile 'org.apache.httpcomponents:httpclient:4.5.5'
-    compile 'commons-io:commons-io:2.6'
-    compile 'org.apache.commons:commons-lang3:3.8.1'
+    compile project(':gateway-connection-api')
     compile 'org.reflections:reflections:0.9.11'
     compile 'com.google.inject:guice:4.2.1'
     compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.9.6'
