+++ gateway-export-plugin/src/main/java/com/ca/apim/gateway/cagatewayexport/util/injection/ExportPluginModule.java
@@ -20,8 +20,8 @@ import com.ca.apim.gateway.cagatewayexport.tasks.explode.linker.EntitiesLinker;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.linker.EntityLinkerRegistry;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.writer.EntityWriter;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.writer.EntityWriterRegistry;
-import com.ca.apim.gateway.cagatewayexport.util.http.GatewayClient;
 import com.ca.apim.gateway.cagatewayexport.util.policy.PolicyXMLSimplifier;
+import com.ca.apim.gateway.connection.GatewayClient;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
