+++ gateway-import-plugin/settings.gradle
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) 2018 CA. All rights reserved.
+ * This software may be modified and distributed under the terms
+ * of the MIT license.  See the LICENSE file for details.
+ */
+
+rootProject.name = 'gateway-import-plugin'
