+++ settings.gradle
@@ -19,4 +19,9 @@ buildscript {
 
 apply plugin: 'net.vivin.gradle-semantic-build-versioning'
 
-include 'config-builder', 'gateway-developer-plugin', 'environment-creator-application', 'gateway-export-plugin'
+include 'config-builder', 
+        'gateway-developer-plugin', 
+        'environment-creator-application', 
+        'gateway-connection-api', 
+        'gateway-export-plugin', 
+        'gateway-import-plugin'
